:sparkles: implement LogRecord to log data with opentelemetry log data model specification
