+++ derive/log/record.py
@@ -3,9 +3,27 @@ import typing
 
 from derive.log.types import ArgsType, SysExcInfoType
 from derive.trace import trace
+import derive
 
 
 class DeriveLogRecord(logging.LogRecord):
+    BUILTIN_RECORD_ATTRS = frozenset(
+        (
+            "exc_info",
+            "exc_text",
+            "filename",
+            "funcName",
+            "lineno",
+            "module",
+            "name",
+            "pathname",
+            "process",
+            "processName",
+            "thread",
+            "threadName",
+            "stack_info",
+        )
+    )
     trace_id: typing.Optional[str]
     attributes: typing.Dict[str, str]
 
@@ -33,3 +51,25 @@ class DeriveLogRecord(logging.LogRecord):
         self.attributes: typing.Mapping[str, str] = (
             {k: str(v) for k, v in extra.items()} if extra else {}
         )
+
+    def to_log_data(self) -> dict:
+        # reference: https://opentelemetry.io/docs/reference/specification/logs/data-model/
+        attributes: typing.Dict[str, str] = {}
+        for attribute in self.BUILTIN_RECORD_ATTRS:
+            attr_value = getattr(self, attribute)
+            if attr_value:
+                attributes[f"builtin_{attribute}"] = str(attr_value)
+
+        data = {
+            "SeverityText": self.levelname,
+            "SeverityNumber": self.levelno,
+            "Body": self.msg,
+            "Timestamp": self.created,
+            "Attributes": {**attributes, **self.attributes},
+            "Resource": {
+                k: str(v) for k, v in derive.get_global_resources().attributes.items()
+            },
+        }
+        if getattr(self, "trace_id") is not None:
+            data["TraceId"] = self.trace_id
+        return data
