+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "derive"
-version = "0.2.0"
+version = "0.2.1"
 description = "The Python Observability Toolkit"
 authors = ["DeBankDeFi <sre@debank.com>"]
